knowledge-graph-python/requirements.txt
neo4j>=5.27.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
